Version compatibility fixes

+++ dask/dataframe/core.py
@@ -1617,7 +1617,7 @@ def _coerce_loc_index(divisions, o):
     if divisions and isinstance(divisions[0], datetime):
         return pd.Timestamp(o)
     if divisions and isinstance(divisions[0], np.datetime64):
-        return np.datetime64(o)
+        return np.datetime64(o).astype(divisions[0].dtype)
     return o
